Modularise routes
