+++ src/routes.js
@@ -1,5 +1,7 @@
 module.exports = [].concat(
   require('./routes/login.js'),
   require('./routes/register.js'),
-  require('./routes/dynamic.js')
+  require('./routes/dynamic_account.js'),
+  require('./routes/dynamic_addreview.js'),
+  require('./routes/dynamic_reviewform.js'),  
 );
