+++ src/routes/dynamic_account.js
@@ -1,5 +1,4 @@
 const getData = require('../getData.js');
-const addReview = require('../addReview.js');
 
 module.exports = [{
   method: 'GET',
@@ -26,40 +25,8 @@ module.exports = [{
         getData.myReviews((error, data) => {
           if (error) throw error;
           userPageData.myReviews = data;
-          reply.view('index', userPageData);
+          return reply.view('index', userPageData);
         }, request.auth.credentials.user);
       }
     }
-},
-{
-  method: 'GET',
-  path:'/account/reviewform/',
-  config: {
-    auth: {
-      strategy: 'base'
-    },
-      handler: (request,reply) => {
-        let userPageData = {
-          user: encodeURIComponent(request.auth.credentials.user),
-          resource: request.query.resource
-        }
-        reply.view('add_review', userPageData);
-      }
-  }
-},
-{
-  method: 'POST',
-  path:'/account/addreview/',
-  config: {
-    auth: {
-      strategy: 'base'
-    },
-      handler: (request,reply) => {
-        let user = encodeURIComponent(request.auth.credentials.user);
-        addReview((error) => {
-          if(error) throw error;
-        },request.payload, user, request.payload.resource_title);
-        reply().redirect('/account/');
-      }
-  }
 }];
